Move SubexpressionStream implementation into interpreter module
