+++ lib/src/interpreter/mod.rs
@@ -1,5 +1,7 @@
 mod plan;
+mod subexpr;
 mod matching;
 
 pub use plan::*;
+pub use subexpr::*;
 pub use matching::interpret;
